Cập nhật Project .NET

+++ Models/Admin/QuanLyMayTinhViewModel.cs
@@ -5,7 +5,7 @@ namespace Project.Models.Admin
         public string? MaMay { get; set; } // Mã máy
         public string? TenMay { get; set; } // Tên máy
         public string? TrangThai { get; set; } // Trạng thái (Đang sử dụng, Sẵn sàng, Bảo trì)
-        public int DonGia { get; set; } // Đơn giá
+        public double DonGia { get; set; } // Đơn giá
         public string? MoTa { get; set; } // Mô tả
     }
 }
